# Refs #28459 -- Improved performance of SQLCompiler.apply_converters().